Add files via upload
The renewed version of "aerial_arm.mlapp" made some revisions so that the change of the system variables of the whole process can be displayed successfully. 
Specifically, I created the arrays "q1_track", "q2_track", "yb_grasp", "psi_grasp", and "fai_grasp" directly in the code. These variables are assumed to not change over time when running the simulation, so they are created as arrays with the same length as time variables,  "time_track" and "time_grasp", and all the elements are the same.
In the previous version, the graphs of the whole process could be displayed in MATLAB R2019a when I worked on this project, but it failed in newer version, such as MATLAB R2023a, so I committed these changes.